+++ android/App/src/main/kotlin/com/darakeon/dfm/extract/MoveAdapter.kt
@@ -7,11 +7,15 @@ import com.darakeon.dfm.lib.ui.Adapter
 class MoveAdapter(
 	activity: ExtractActivity,
 	list: List<Move>,
-	private val canCheck: Boolean
+	private val canCheck: Boolean,
+	private val edit: (MoveLine) -> Unit,
+	private val delete: (MoveLine) -> Unit,
+	private val check: (MoveLine) -> Unit,
+	private val uncheck: (MoveLine) -> Unit,
 ) : Adapter<ExtractActivity, Move, MoveLine>(activity, list) {
 	override val lineLayoutId: Int
-		get() = R.layout.move_details
+		get() = R.layout.move_line
 
 	override fun populateView(view: MoveLine, position: Int) =
-		view.setMove(list[position], canCheck)
+		view.setMove(list[position], canCheck, edit, delete, check, uncheck)
 }