+++ android/App/src/main/kotlin/com/darakeon/dfm/extract/MoveLine.kt
@@ -4,8 +4,6 @@ import android.content.Context
 import android.content.res.Configuration.ORIENTATION_LANDSCAPE
 import android.content.res.Configuration.ORIENTATION_PORTRAIT
 import android.util.AttributeSet
-import android.view.Menu
-import android.view.View
 import android.widget.LinearLayout
 import com.darakeon.dfm.R
 import com.darakeon.dfm.lib.api.entities.extract.Move
@@ -13,25 +11,35 @@ import com.darakeon.dfm.lib.api.entities.moves.Nature
 import com.darakeon.dfm.lib.extensions.applyGlyphicon
 import com.darakeon.dfm.lib.extensions.setColorByAttr
 import com.darakeon.dfm.lib.extensions.setValueColored
-import kotlinx.android.synthetic.main.move_details.view.move_status
 import kotlinx.android.synthetic.main.move_details.view.date
+import kotlinx.android.synthetic.main.move_details.view.move_status
 import kotlinx.android.synthetic.main.move_details.view.name
 import kotlinx.android.synthetic.main.move_details.view.value
+import kotlinx.android.synthetic.main.move_line.view.action_check
+import kotlinx.android.synthetic.main.move_line.view.action_delete
+import kotlinx.android.synthetic.main.move_line.view.action_edit
+import kotlinx.android.synthetic.main.move_line.view.action_uncheck
+import kotlinx.android.synthetic.main.move_line.view.actions
 import java.util.UUID
 
 class MoveLine(
 	context: Context,
 	attributeSet: AttributeSet
 ) : LinearLayout(context, attributeSet) {
-	var menu: Menu? = null
-
 	private lateinit var move: Move
 	private var checkNature: Nature = Nature.Out
 
 	val guid: UUID
 		get() = move.guid
 
-	fun setMove(move: Move, canCheck: Boolean) {
+	fun setMove(
+		move: Move,
+		canCheck: Boolean,
+		edit: (MoveLine) -> Unit,
+		delete: (MoveLine) -> Unit,
+		check: (MoveLine) -> Unit,
+		uncheck: (MoveLine) -> Unit,
+	) {
 		this.move = move
 
 		name.text = move.description
@@ -40,8 +48,14 @@ class MoveLine(
 		setCheckNature(move)
 		setDateField(move)
 
+		setActions(edit, delete, check, uncheck)
+
 		setOnLongClickListener {
-			it.showContextMenu()
+			if (actions.visibility == GONE)
+				actions.visibility = VISIBLE
+			else
+				actions.visibility = GONE
+
 			true
 		}
 	}
@@ -54,7 +68,9 @@ class MoveLine(
 		if (canCheck) {
 			setCheckField()
 		} else {
-			move_status.visibility = View.GONE
+			action_check.visibility = GONE
+			action_uncheck.visibility = GONE
+			move_status.visibility = GONE
 		}
 	}
 
@@ -66,6 +82,15 @@ class MoveLine(
 
 		val color = if (isChecked) R.attr.checked else R.attr.unchecked
 		move_status.setColorByAttr(color)
+
+		if (isChecked) {
+			action_check.visibility = GONE
+			action_uncheck.visibility = VISIBLE
+		}
+		else {
+			action_check.visibility = VISIBLE
+			action_uncheck.visibility = GONE
+		}
 	}
 
 	private fun setDateField(move: Move) {
@@ -86,6 +111,25 @@ class MoveLine(
 				Nature.Out
 			else
 				Nature.In
+	}
+
+	private fun setActions(
+		edit: (MoveLine) -> Unit,
+		delete: (MoveLine) -> Unit,
+		check: (MoveLine) -> Unit,
+		uncheck: (MoveLine) -> Unit,
+	) {
+		action_edit.applyGlyphicon()
+		action_edit.setOnClickListener { edit(this) }
+
+		action_delete.applyGlyphicon()
+		action_delete.setOnClickListener { delete(this) }
+
+		action_check.applyGlyphicon()
+		action_check.setOnClickListener { check(this) }
+
+		action_uncheck.applyGlyphicon()
+		action_uncheck.setOnClickListener { uncheck(this) }
 	}
 
 	fun check() {